+++ src/main/java/com/spring/miniposbackend/exception/ExceptionResponse.java
@@ -8,11 +8,11 @@ import lombok.Getter;
 public class ExceptionResponse  {
 
 	private Date timestamp;
-	private Integer code;
+	private Short code;
 	private String message;
 	private String path;
 	
-	public ExceptionResponse(Date timestamp, Integer code, String message) {
+	public ExceptionResponse(Date timestamp, Short code, String message) {
 		this.timestamp = timestamp;
 		this.code = code;
 		this.message = message;
